refactor: split the business functions into multiple files
